requirements.txt
polygon-api-client>=1.14.0
pymongo>=4.6.0
motor>=3.3.0
pyspark>=3.5.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
